configure: Add support for external METIS library
Let the user tell us at configure-time that METIS is available via
--with-metis. If they don't specify a directory, we assume the system
paths include it. If they specify a directory, look for libmetis.a in
exactly that directory, and find metis.h relative to it based on how
METIS 4 and 5 are laid out.

When METIS is available, then we build MetisLB and TeamLB as part of
the charm++ target.