Distributed Completion Detection: Apps don't need to use QD
A mechanism to detect the completion of distributed processes, based
on a counts of event generators and events produced and consumed. From
an application's standpoint, this accomplishes the same thing as QD,
but on a more modular basis.

Ideally, this should be able to prioritize itself below the process
it's tracking. That's not implemented yet.