Print small sizes as ints, to avoid size_t's printf portability morass
Calling printf("%d", sizeof(Foo)) generates warnings, because size_t
is often not an int. We want to squelch all the warnings we're seeing,
but unfortunately, the standard %zd size modifier isn't available on
windows. Since Foo in this case is the small struct _Slot, we don't
need to worry about overflow, and can just shorten its size to int.