ckarrayindex: Return combined count only for dense (upto 3D) arrays
Code for the creation of dense multi-dimensional chare arrays uses an array
index object to store the dimensions of the array.

The index object sports a funny method introduced to support such use. Ensure
that it returns a count on the number of array elements only for arrays upto 3D
(which are created dense by default).