array index: Introduce a union-friendly base class for ckarrayindex
There are only three entities in the charm codebase that put an array
index within a union:
    - envelope
    - callbacks
    - ckobjid

Create a base class that is used just by these three entities. CkArrayIndex is
now a safe and usable index class as it sports constructors, initialization
etc.  The remainder of the codebase and charm applications can and should use
the CkArrayIndex class.

Just at locations where the index objects in unions are handled, we provide a
somewhat "risky" cast operation from base to derived. This should be less
insane than casting across unrelated classes, which was the root of the
perennial problems with the original CkArrayIndex, IndexMax and IndexStruct
setup. However, we should actively explore mechanisms whereby even such casts
can be avoided without too much time/space costs.