LBPredictor: Fix old re-initialization code that would lead to crashes
The code to change from one load predictor to another was overwriting the
validity array pointer with 'false', effectively NULL-ing it out. This would
cause a crash on the next use, but was apparently never tested. That code
was supposed to make each entry in the array false, so do that instead.

Detected thanks to g++-4.7 -Werror.