charmxi: Restore original expectations on entry method argument visibility in decl.h
Recent changes exposed applications that depended on charm's relaxed (lack
of) requirements on the visibility of entry method argument declarations and
definitions in the module's decl.h generated header. This fix restores this
original relaxed requirement by moving the actual ep registration code back
into the def.h file. This should unbreak the applications.

This still retains the idx_foo() methods in the decl.h so that they can be
inlined by the compiler and templated by the code generator in the future.