msgQ: simplify enq and deq by eliminating unnecessary vector of msg buckets
The original intent of a vector of msg buckets was to permit constant-time deq()
and to permit features such as a randomized message queue. The implementation
now eschews this additional container in favor of simply storing pointers to the
buckets directly. The constant-time deq() is still retained, while randomized queues
will have to jump through a few additional hoops.