Fix mis-use of inline in C
This is breaking on Solaris because it has a proper C99 implementation
of the inline keyword, which is different from its meaning in C++.
You can make a static function inline, but if you make a non-static
function inline then exactly one compilation unit also has to declare
the function "extern inline" to avoid having the symbol undefined.

See:

http://clang.llvm.org/compatibility.html

http://gustedt.wordpress.com/2010/11/29/myth-and-reality-about-inline-in-c99/

http://www.drdobbs.com/the-new-c-inline-functions/184401540

All of the other places where CMK_C_INLINE is used it is only adding
the inline keyword to an already static function.