Avoid undefined behavior of left-shifting through the sign bit
In C and C++, shifting left into or past the sign bit of a signed
integer results in undefined behavior. In cases where we're just
setting up a bit mask, use the bitwise inverse of an unsigned 0
instead, on which shifting is perfectly well defined.

Detected with
clang -fsanitize=integer -fno-sanitize=unsigned-integer-overflow