CkHashCode: return unmodified when rotating by 0 to avoid undefined behavior
Shifting by more the full width of a variable is undefined behavior in
C and C++. When trying to rotate by 0, a shift of the full width
resulted. Avoid that by returning the unmodified argument when no
rotation is to be applied. The test should be free, either because the
value was just computed and so the condition code is set, or because
the compiler emits a rotate opcode anyway.

Error detected by
clang -fsanitize=integer -fno-sanitize=unsigned-integer-overflow

This leaves the entire execution of tests/charm++ clean of undefined
signed integer overflow.