rednmgr: remove derived class overload of virtual flushStates() that hid base class method
global and local element counters in CkReductionMgr need to be reset only when
the redn mgr is serving a chare array (and not when its serving a group). Implementing
this via an input flag to flushStates() caused flushStates() to become overloaded in
the child class. This overload hid the base class virtual method causing some compilers
(icc?) to complain (warnings). This somewhat cleaner solution should avoid the
warnings, but it still stinks of a hack.