Revert "Add appropriate flags for charmc -optimize when using Intel compilers"
It turns out the -fast flag to icc forces static linking, which is unacceptable
for various reasons (off the top of my head, broken memory allocator choice,
incompatibility with some external libraries, loss of some options for global
privatization in AMPI). The default for -optimize of -O2 should be good enough.

This reverts commit 852b02387d887802a04506d9d33525634c5afaf3.