Clarify documentation for CkReduction::set
The current description is missing important use information. It is not clear
how to pass the data into the reduction in the first place. The example given
is incorrect. The data field on CkReductionMsg is private, it has to be
accessed through the getData() getter method. I have added example code for
the reduction method and for the reductiontarget method.