Fixed bug in CkReductionMgr::reduceMessages().
The first contribution of a child node is a single message. In this case
CkReductionMgr::reduceMessages() simply returns the message without calling
the reducer. However, when using the CkReduction::set reducer this behavior is
incorrect, and instead the reducer should be called even for one single
message so that the message is wrapped into a setElement struct. In the
current implementation, the reduction becomes corrupted because the
CkReduction::set() method one tier up in the reduction sizes the remote
contribution incorrectly assuming a setElement struct and not a simple
message.