Enforce the invariant that currentChareIdx is only valid when a plain chare is being constructed
Fix a bug where the currentChareIdx is set to idx and not reset. The bug
manifests in this simple example where a single chare is created followed by
dynamic insertion of array elements. This sets the value of chareIdx of the
array element to be the id of the single chare and erroneously deregisters that
chare upon deletion of the array element.

Change-Id: I6a2401f81e7de67e7e4faa1ab9810cad467ee108