Ensure that currentChareIdx is -1 when constructing Group and Array elements
Fix a bug where the currentChareIdx is set to idx and not reset. The bug
manifests in this simple example where a single chare is created followed by
dynamic insertion of array elements. This sets the value of chareIdx of the
array element to be the id of the single chare and erroneously deregisters that
chare upon deletion of the array element.

Second attempt at what's implemented in commit
6d3385f798cb8ee1b81c41808a562bf1d348886d

Change-Id: I48ab549fa7a79de1e009e6b48524257afa543fc7