PUP: Define overloads for PUPbytes types in "namespace PUP"
We want to ensure that the overloads of operator| and PUParray can always be
found, no matter where they're referenced. Moving them into the PUP namespace
gets us there, because the PUP::er argument is of a type defined in that
namespace.

As written, this requires that PUPbytes(Foo) declarations be written *outside*
any specific namespace, in the global namespace. Otherwise, it will result in
defining a namespace PUP inside that other namespace.

Change-Id: Idb6deb8803258da39c51ab84934fe4c539a5b291