Add asynchronous array construction API for post-startup use cases
When a chare array is created using this API, the callback function
receives a message, m, that contains the array id (m->aid).

This is necessary to provide positive array manager group IDs.

In future, when 64 bit ID is adopted, we want the group IDs to be
limited to a smaller compact range, hence the need to create them on
PE0. This API will enable us to force that requirement.

Change-Id: I6c22dab40199374cd18e85a75a166713d3e90bb6