Bug #304: Some libraries are only built statically, preventing installation with only dynamic libraries
On a typical Linux system only shared libraries are installed, because of the
usual security and bloat arguments. However, if I build charm with
--build-shared, and then install only what is in lib_so, the charmc script
breaks because some Converse libraries are built purely as .o and don't end up
in lib_so. One example is seed based load balancing, e.g. libldb-rand.o. As far
as I can tell, this is not too hard to fix. In the main makefile, the targets
libldb-*.o are renamed to libldb-*.a, and the charmc script now tests both .a
and .so when linking the seed based load balancing code.

Change-Id: I52167e0736950bdadd46c6c320dec4024bbb6c8d