Disable stack-smashing protection based on relevant compilers, not OS
Stack-smashing protection is implemented on Mac OS and Windows, in addition to
Linux, by any or all of GCC, ICC, and Clang. It's been observed to cause
consistent crashes on Mac OS, which disppear when it's disabled. So, broaden
the efforts to disable it automatically, since it doesn't serve the purposes of
Charm++ code.

Change-Id: I9d47fde5e3ca3ae4241dc2233dc6816c000320da