Bug #705: Group nodelist entries by line number
Before this commit, nodes were added to the nodetab list as the nodelist
file was parsed. This meant that in the wrap-around phase, nodes
corresponding to the same host might not be consecutive, screwing up the
naive scalable-start scheme. This change bins the nodes by line number
before inserting them in the node table, enabling scalable-start even
when wrap-around occurs.

Change-Id: I5035812b7cebb26c312f52144a441568a45775d5