Bug #713: Fix tsan+smp deadlock
The combination of tsan+smp deadlocks inside a 'write' syscall that is
(by default) writing to stderr, which has been replaced by a pipe, on
the reading side of which is the commthread. It turned out that the
commthread was blocking on a mutex, inside which tsan tries to report
something, leading to a deadlock. In this commit we prohibit TSAN from
writing to standard error, aborting if the user does not specify a log
file to write to (using TSAN_OPTIONS.)

Change-Id: I924bb9a9fc13e0008987ddbc0513649960d6ff65