Bug 890 charmrun: Fix skt_to_name to return right hostname
skt_to_name was returning wrong hostname associated with a socket.
skt_to_name uses skt_client_table which is supposed to map a socket to the node id
using the socket, but skt_client_table contained ids based on the order on which connections
were received, not to the node id of the actual client.

The fix corrects skt_client_table by associating a socket with the node id
using that socket.

Change-Id: I682d5a0b1d96d67e11cb87d75d9479384ff9a611