Bug #788: enable cuda out of tree build for all layers
This propogrates the change in b4a496 to work
everywhere. It also partially adresses Cleanup #512
since we are removing duplicated special.sh and
keeping them in a single file in common.
net-linux-cell still has its own version of
special.sh and that should overwrite the common
version during the gather stage due to system_ln
passing the -f flag.

Change-Id: I21ee6cd62ece09e225bdf1147e6155d1a608391f