LRTS locking API
Adding a new locking API to LRTS. The goal of the API is to expose the machine
specific locks to charm via the common LRTS interface.
Default code provides locking with pthread mutexes and CMK_USE_COMMON_LOCK
macro needs to be enabled for the use of default common lock.

Change-Id: Ie4d616094c6758b36b8e933d3027e7875f1c501c