Bug #917 follow-up: Change the test configurations for chkpt example
With commit ca884ee0, autobuild is skipping the tests if ppn does not
divide p. This creates a problem in checkpoint example where the
application tries to restart from the logs with a different p number
that divides ppn but checkpoint logs are not there because the run that
supposed to create them is skipped.

Change-Id: I672d413ea41ac247182e1519685052e90fc2d712