AMPIF #1078: Terminate Fortran strings correctly.
Strings passed from Fortran into AMPI need to be NULL terminated.
We now define 'MPI_' routines that take string arguments in
ampifimpl.f90, where we get the length of the string in Fortran and pass
that into the wrappers in ampif.C. In ampif.C, we explicitly NULL
terminate the strings before passing them into 'AMPI_' routines.

Strings passed to Fortran from AMPI need to be backfilled with spaces.
In ampif.C, we now copy strings that will be passed back to the user
after calls into 'AMPI_' routines and fill the end of the string with
spaces.

Change-Id: I22c8e44bf9bdf28f902cba2c2c223b31f0fa011d