AMPI #1111: avoid sender-side copy for large contiguous messages using Charm++'s zero copy RDMA send API
For large contiguous messages, we now use rdma entry methods and track
completion of the send buffer via a new request type, SendReq.

A sender invokes genericRdma with a callback to completedRdmaSend to
complete the send request. A receiver gets the rdma'ed message as if it
were parameter marshalled, so if the message is expected/in-order we
deserialize directly from there into the user's buffer. If the message
is unexpected/out-of-order, we copy it to an AmpiMsg and stash it in a
message queue.

The new control flow for receivers is genericRdma -> inorderRdma ->
receiveRdma -> processRdmaMsg. These functions all have non-Rdma
equivalents, the difference is in dealing with AmpiMsgs vs. raw buffers.

Change-Id: Id0867af29bb17ac4abd9d6ac340735870885ddaa