Bug #1045 follow-up: recognize IBM XL compilers's idiosyncratic warning format
This prevents the configure script from incorrectly passing the
GCC-specific flag -fno-lifetime-dse to a compiler that doesn't
recognize it.

Note that the configure output will still say "yes" for that test,
because that output is based on whether the compiler errored out or
not. The script correctly omits the flag from future tests and its
resulting settings.

Change-Id: Ib71b786696aec967e6561338f4b05ed03f7fb5d6