CkMulticast: Tree creation performance optimization
Changes can substantially speed up the generation of
spanning trees (I have seen up to 2.4x in my tests).

Main change is that, instead of storing the list
of section elements and their last known PE in the
setup message, I store the list of section elements,
and the list of PEs in the subtree (and for each PE
I include a "pointer" to its elements in the element list).
This is done at the root, and is carried down the tree,
avoiding having to recalculate which element belongs
to which PE on every tree node, and avoids useless
copying of array indexes.

Change-Id: Ia8e3713e7eca503fdc775dd310198df1ad096d25