Feature #1458: Zero-copy send support for the MPI machine layer
Support for MPI machine layer was added using MPI_Isend and MPI_Irecv.
A new tag is generated for every buffer that is to be sent using the
zero copy send API. For both the Converse and the LRTS layer, the methods
CmiSetRdmaRecvInfo and LrtsSetRdmaRecvInfo  of the Rdma Converse/LRTS API
were modified to take an additional integer input parameter 'msgSize' that
represents the size of the new message containing both the metadata message
and the receiver allocated buffer. For the MPI layer, this was required to
be explicitly set by the charm layer, because unlike other layers, the msgSize
is not stored as a part of the converse header. The msgSize is used in calling
the method 'handleOneRecvedMsg' which calls the Converse message handler.

Change-Id: If55921cd7b408cfbf93682fe2842a07a7a6a85e2