Bug #664/Bug #1473: Disable test that fails due to its own race condition
With an assertion added that the messages received match the size of
messages currently expected, tests/charm++/communication_overhead
crashes. A larger message from a later iteration can end up being
delivered before a smaller message at the end of the preceding
iteration.

Disable this broken test for now, so that we can see that everything
else is working and get a release out the door.

Change-Id: I30b6b0a393bd9a0a9d0e1d347aa1576406aa4a1d