Bug #1537: fixing the bug that breaks shrink/expand
Earlier commit that replaces signal with sigaction causes a hang
in shrink/expand. signal function only handles one instance of the
signal and returns to default behavior afterwards. To be able to get the
same behavior with sigaction, I added a handler function which returns
to default behavior after the first signal.

Change-Id: I415a2499746bb929f3c16f61a65ed357d06b0b95